Minor code restructuring in recipes and models
