+++ src/fairseq2/datasets/loader.py
@@ -70,6 +70,7 @@ class AbstractDatasetLoader(ABC, DatasetLoader[DatasetT]):
         self._asset_store = asset_store or default_asset_store
         self._download_manager = download_manager or default_download_manager
 
+    @final
     def __call__(
         self,
         dataset_name_or_card: Union[str, AssetCard, Path],
