+++ src/fairseq2/models/mistral/archs.py
@@ -12,9 +12,6 @@ mistral_archs = ConfigRegistry[MistralConfig]()
 mistral_arch = mistral_archs.decorator
 
 
+@mistral_arch("7b")
 def _7b() -> MistralConfig:
     return MistralConfig()
-
-
-def _register_mistral_archs() -> None:
-    mistral_archs.register("7b", _7b)
