+++ src/fairseq2/models/nllb/archs.py
@@ -5,10 +5,15 @@
 # LICENSE file in the root directory of this source tree.
 
 from fairseq2.data import VocabularyInfo
-from fairseq2.models.transformer import TransformerConfig, transformer_archs
+from fairseq2.models.transformer import (
+    TransformerConfig,
+    transformer_arch,
+    transformer_archs,
+)
 from fairseq2.nn.transformer import TransformerNormOrder
 
 
+@transformer_arch("nllb_dense_300m")
 def _dense_300m() -> TransformerConfig:
     config = _dense_1b()
 
@@ -19,6 +24,7 @@ def _dense_300m() -> TransformerConfig:
     return config
 
 
+@transformer_arch("nllb_dense_600m")
 def _dense_600m() -> TransformerConfig:
     config = _dense_1b()
 
@@ -29,6 +35,7 @@ def _dense_600m() -> TransformerConfig:
     return config
 
 
+@transformer_arch("nllb_dense_1b")
 def _dense_1b() -> TransformerConfig:
     config = transformer_archs.get("base")
 
@@ -46,18 +53,10 @@ def _dense_1b() -> TransformerConfig:
     return config
 
 
+@transformer_arch("nllb_dense_3b")
 def _dense_3b() -> TransformerConfig:
     config = _dense_1b()
 
     config.model_dim = 2048
 
     return config
-
-
-def _register_nllb_archs() -> None:
-    # fmt: off
-    transformer_archs.register("nllb_dense_300m", _dense_300m)
-    transformer_archs.register("nllb_dense_600m", _dense_600m)
-    transformer_archs.register("nllb_dense_1b",   _dense_1b)
-    transformer_archs.register("nllb_dense_3b",   _dense_3b)
-    # fmt: on
