+++ src/fairseq2/models/wav2vec2/asr/__init__.py
@@ -24,13 +24,3 @@ from fairseq2.models.wav2vec2.asr.loader import (
 )
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrModel as Wav2Vec2AsrModel
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrOutput as Wav2Vec2AsrOutput
-
-# isort: split
-
-from fairseq2.models.wav2vec2.asr.archs import _register_wav2vec2_asr_archs
-from fairseq2.models.wav2vec2.asr.loader import _register_wav2vec2_asr_loaders
-
-
-def _register_wav2vec2_asr() -> None:
-    _register_wav2vec2_asr_archs()
-    _register_wav2vec2_asr_loaders()
