+++ src/fairseq2/recipes/llama/__init__.py
@@ -5,14 +5,14 @@
 # LICENSE file in the root directory of this source tree.
 
 from fairseq2.recipes.cli import Cli
-from fairseq2.recipes.llama.convert import ConvertCheckpointCommandHandler
+from fairseq2.recipes.llama.convert_checkpoint import ConvertCheckpointCommandHandler
 
 
 def _setup_llama_cli(cli: Cli) -> None:
     group = cli.add_group("llama", help="LLaMA recipes")
 
     group.add_command(
-        name="convert",
+        name="convert_checkpoint",
         handler=ConvertCheckpointCommandHandler(),
         help="convert fairseq2 LLaMA checkpoints to reference checkpoints",
     )
